#237 [REFACTOR] 통계와 관련된 API 를 /project/ 에서 /statistics/ 로 위치 변경
