+++ src/main/java/com/developlife/reviewtwits/controller/StatController.java
@@ -1,17 +1,19 @@
 package com.developlife.reviewtwits.controller;
 
 import com.developlife.reviewtwits.entity.User;
+import com.developlife.reviewtwits.message.annotation.project.ChartPeriod;
 import com.developlife.reviewtwits.message.request.StatMessageRequest;
+import com.developlife.reviewtwits.message.response.project.DailyVisitInfoResponse;
+import com.developlife.reviewtwits.message.response.project.RecentVisitInfoResponse;
+import com.developlife.reviewtwits.message.response.project.VisitTotalGraphResponse;
 import com.developlife.reviewtwits.message.response.statistics.SaveStatResponse;
 import com.developlife.reviewtwits.service.StatService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import javax.validation.constraints.Min;
 
 /**
  * @author WhalesBob
@@ -30,5 +32,28 @@ public class StatController {
                                             @RequestBody @Valid StatMessageRequest statMessageRequest){
 
         return statService.saveStatInfo(user,statMessageRequest);
+    }
+
+    @GetMapping("/visit-graph-infos")
+    public VisitTotalGraphResponse getVisitGraphInfos(@AuthenticationPrincipal User user,
+                                                      @RequestParam
+                                                      @Min(value = 1, message = "프로젝트 아이디는 1 이상의 수로 입력해야 합니다.") Long projectId,
+                                                      @RequestParam @ChartPeriod String range,
+                                                      @RequestParam @ChartPeriod String interval){
+
+        return statService.getVisitGraphInfos(projectId, range, interval, user);
+    }
+    @GetMapping("/daily-visit-graph-infos")
+    public DailyVisitInfoResponse getDailyVisitInfos(@AuthenticationPrincipal User user,
+                                                     @RequestParam
+                                                     @Min(value = 1, message = "프로젝트 아이디는 1 이상의 수로 입력해야 합니다.") Long projectId,
+                                                     @RequestParam @ChartPeriod String range){
+        return statService.getDailyVisitInfos(projectId, range, user);
+    }
+    @GetMapping("/recent-visit-counts")
+    public RecentVisitInfoResponse getRecentVisitCounts(@AuthenticationPrincipal User user,
+                                                        @RequestParam @Min(value = 1, message = "프로젝트 아이디는 1 이상의 수로 입력해야 합니다.") Long projectId
+    ){
+        return statService.getRecentVisitCounts(projectId, user);
     }
 }
